+++ projecten3-1920-backend-klim03/Domain/Models/DTOs/CustomDTOs/PdfSettings.cs
@@ -7,7 +7,7 @@ namespace projecten3_1920_backend_klim03.Domain.Models.DTOs.CustomDTOs
 {
     public class PdfSettings
     {
-        public List<int> GroupsToShow { get; set; }
+        public List<long> GroupsToShow { get; set; }
 
         public bool ShowPupil { get; set; }
         public bool ShowTeacher { get; set; }
